@@ -10,14 +10,13 @@ Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
-limitations under the License. */ 
+limitations under the License. */
 
 import { Injectable } from '@angular/core';
 import { Http, Response, Headers, RequestOptions } from '@angular/http';
-import { Observable } from 'rxjs/Observable';
-import { of } from 'rxjs/observable/of';
+import { Observable } from 'rxjs';
 import { Report } from './report';
-import { catchError, map, tap } from 'rxjs/operators';
+import { map } from 'rxjs/operators';
 
 @Injectable()
 export class ReportsService {
@@ -35,7 +34,9 @@ export class ReportsService {
 
   checkAccess(): Observable<string> {
     return this.http.get(this.baseUrl + this.checkAccessUrl)
-             .map((res:Response) => res.json().role);
+             .pipe(
+                 map((res:Response) => res.json().role)
+             );
   }
 
   getReports(body: Object): Observable<Report[]> {
@@ -45,20 +46,26 @@ export class ReportsService {
     let options = new RequestOptions({ headers: headers });
 
     return this.http.post(this.baseUrl + this.reportListUrl, body, options)
-             .map((res:Response) => res.json().reports);
+             .pipe(
+                 map((res:Response) => res.json().reports)
+             );
 
   }
 
   getOrgs(): Observable<string[]> {
 
     return this.http.get(this.baseUrl + this.orgListUrl)
-             .map((res:Response) => res.json().orgs);
+             .pipe(
+                 map((res:Response) => res.json().orgs)
+             );
   }
 
   getDatasources(): Observable<string[]> {
 
     return this.http.get(this.baseUrl + this.datasourceListUrl)
-             .map((res:Response) => res.json().tables);
+             .pipe(
+                 map((res:Response) => res.json().tables)
+             );
   }
 
   addReport(body: Object): Observable<Report[]> {
@@ -68,7 +75,9 @@ export class ReportsService {
     let options = new RequestOptions({ headers: headers });
 
     return this.http.post(this.baseUrl + this.addReportUrl, body, options)
-             .map((res:Response) => res.json().reports);
+             .pipe(
+                 map((res:Response) => res.json().reports)
+             );
   }
 
   deleteReport(body: Object): Observable<Report[]> {
@@ -77,7 +86,9 @@ export class ReportsService {
     let options = new RequestOptions({ headers: headers });
 
     return this.http.post(this.baseUrl + this.deleteReportUrl, body, options)
-             .map((res:Response) => res.json().reports);
+             .pipe(
+                 map((res:Response) => res.json().reports)
+             );
   }
 
   editReport(body: Object): Observable<Report[]> {
@@ -86,7 +97,9 @@ export class ReportsService {
     let options = new RequestOptions({ headers: headers });
 
     return this.http.post(this.baseUrl + this.editReportUrl, body, options)
-             .map((res:Response) => res.json().reports);
+             .pipe(
+                 map((res:Response) => res.json().reports)
+             );
   }
 
 }
